Reorganisation of processes to separate main runner from process_evaluation
