+++ MetricsReloaded/metrics/prob_pairwise_measures.py
@@ -52,8 +52,6 @@ class ProbabilityPairwiseMeasures(object):
         ref_proba,
         case=None,
         measures=[],
-        num_neighbors=8,
-        pixdim=[1, 1, 1],
         empty=False,
         dict_args={},
     ):
